lambda/retraining_data_labeling/requirements.txt
sqlalchemy
psycopg2-binary
boto3
pydantic
google-genai
python-dotenv
exceptiongroup
